// packages/resources/version.ts
import { Resource, ResourceTemplate } from "@c6o/kubeclient-contracts";

/** Info contains versioning information. how we'll want to distribute that information. */
export interface Info {
    buildDate: string;
    compiler: string;
    gitCommit: string;
    gitTreeState: string;
    gitVersion: string;
    goVersion: string;
    major: string;
    minor: string;
    platform: string;
}